+++ src/app/admin-interests/admin-interests.component.ts
@@ -21,7 +21,7 @@ export class AdminInterestsComponent {
     hobbiesActividades: '',
     interesesProfesionales: ''
   };
-  
+
   interests?: Interest[];
 
   constructor(public interestService: InterestService) {
@@ -40,7 +40,29 @@ export class AdminInterestsComponent {
       });
   }
 
-  // Método para agregar o actualizar un interés
+  AgregarInterest() {
+    this.interestService.createInterest(this.myInterest).then(() => {
+      this.resetForm();
+    });
+  }
+
+  EditarInterest(interest: Interest) {
+    this.isEditing = true;
+    this.editingId = interest.id || null;
+    this.myInterest = { ...interest };
+    this.btnTxt = 'Actualizar';
+  }
+
+  ActualizarInterest() {
+    if (!this.editingId) return;
+    const confirmacion = window.confirm('¿Estás seguro de guardar los cambios?');
+    if (confirmacion) {
+      this.interestService.updateInterest(this.editingId, this.myInterest).then(() => {
+        this.resetForm();
+      });
+    }
+  }
+
   AgregarOActualizar() {
     if (this.isEditing) {
       this.ActualizarInterest();
@@ -49,38 +71,16 @@ export class AdminInterestsComponent {
     }
   }
 
-  // Método para agregar un interés
-  AgregarInterest() {
-    this.interestService.createInterest(this.myInterest).then(() => {
-      this.resetForm();
-    });
-  }
-
-  // Método para editar un interés
-  EditarInterest(interest: Interest) {
-    this.isEditing = true;
-    this.editingId = interest.id || null;
-    this.myInterest = { ...interest };
-    this.btnTxt = 'Actualizar';
-  }
-
-  // Método para actualizar un interés
-  ActualizarInterest() {
-    if (!this.editingId) return;
-    this.interestService.updateInterest(this.editingId, this.myInterest).then(() => {
-      this.resetForm();
-    });
-  }
-
-  // Método para eliminar un interés
   deleteInterest(id?: string) {
     if (!id) return;
-    this.interestService.deleteInterest(id).then(() => {
-      console.log('Interés eliminado correctamente!');
-    });
+    const confirmacion = window.confirm('¿Estás seguro de eliminar este interés?');
+    if (confirmacion) {
+      this.interestService.deleteInterest(id).then(() => {
+        console.log('Interés eliminado correctamente!');
+      });
+    }
   }
 
-  // Método para resetear el formulario
   resetForm() {
     this.myInterest = {
       actividadesExtracurriculares: '',
